started project
code to train neural network and compute the adjacency matrix. someone should look at eigenmaps.ipynb and compute the weight matrix and from there the laplacian. there is an outstanding question of what to do with the biases - ignore them entirely or incorporate them in to the weight matrix in some meaningful way.